# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "法律智能问答系统 API",
        "title": "LexVeritas API",
        "contact": {
            "name": "LexVeritas nsxzhou",
            "email": "1790146932@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页查询用户列表,支持按邮箱、角色、状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "邮箱筛选",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "user",
                            "admin",
                            "super_admin"
                        ],
                        "type": "string",
                        "description": "角色筛选",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive",
                            "banned"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词搜索(邮箱或姓名)",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据用户ID获取详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "软删除指定用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/quota": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "调整指定用户的Token配额",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "调整用户Token配额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配额调整请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.AdjustQuotaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "调整成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/role": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "修改用户角色,需要super_admin权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "修改用户角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/status": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "修改用户的激活状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "修改用户状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "使用邮箱和密码登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "邮箱或密码错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "账户被禁用或锁定",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/login/phone": {
            "post": {
                "description": "使用手机号和验证码登录系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "手机验证码登录",
                "parameters": [
                    {
                        "description": "手机登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.PhoneLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "验证码错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "登出当前用户，使访问令牌失效",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前登录用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取当前用户",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/auth/oauth/{provider}": {
            "get": {
                "description": "重定向到第三方OAuth授权页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "OAuth 第三方登录",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github",
                            "wechat"
                        ],
                        "type": "string",
                        "description": "OAuth提供商",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到OAuth授权页面"
                    },
                    "501": {
                        "description": "OAuth登录尚未实现",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.TokenPair"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "刷新令牌无效或已过期",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_lexveritas_lex-veritas-backend_internal_dto.AdjustQuotaRequest": {
            "type": "object",
            "required": [
                "newQuota",
                "userId"
            ],
            "properties": {
                "newQuota": {
                    "type": "integer",
                    "minimum": 0
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.UserListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UserResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}